bringing the sql files up to date for the 2014 data. also missed some files in the last commit from merge with other code base.

+++ time-standards-xcode/Classes/STSTimeStandardAndSwimmerController.m
@@ -431,8 +431,8 @@
     NSUInteger section = indexPath.section;
     switch (section) {
         case timeStandardSection: {
-            int newRow = [indexPath row];
-            int oldRow = (_lastIndexPath != nil) ? [_lastIndexPath row] : -1;
+            NSInteger newRow = [indexPath row];
+            NSInteger oldRow = (_lastIndexPath != nil) ? [_lastIndexPath row] : -1;
             
             if (newRow != oldRow) {
                 UITableViewCell * newCell = [tableViewParam cellForRowAtIndexPath:indexPath];
